# internal/handlers/auth.go
package handlers

import (
	"database/sql"
	"encoding/json"
	"errors"
	"net/http"
	"time"

	"github.com/golang-jwt/jwt/v5"
	"golang.org/x/crypto/bcrypt"
)

var jwtSecretKey = []byte("super-tajny-klucz-inzynierski")

type AuthHandler struct {
	DB *sql.DB
}

type LoginRequest struct {
	Username string `json:"username"`
	Password string `json:"password"`
}

// Login obsługuje uwierzytelnianie administratora
// @Summary      Logowanie do systemu
// @Description  Weryfikuje poświadczenia i zwraca token JWT ważny 24 godziny.
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        request body handlers.LoginRequest true "Dane logowania (admin / admin123)"
// @Success      200  {object}  map[string]string "Zwraca token JWT"
// @Failure      400  {string}  string "Niepoprawny format JSON"
// @Failure      401  {string}  string "Nieprawidłowe dane logowania"
// @Router       /auth/login [post]
func (h *AuthHandler) Login(w http.ResponseWriter, r *http.Request) {
	var req LoginRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Niepoprawny format żądania JSON", http.StatusBadRequest)
		return
	}

	var storedHash string
	err := h.DB.QueryRow(`SELECT password_hash FROM users WHERE username = $1`, req.Username).Scan(&storedHash)
	if err != nil {
		if errors.Is(err, sql.ErrNoRows) {
			http.Error(w, "Nieprawidłowe dane logowania", http.StatusUnauthorized)
		} else {
			http.Error(w, "Błąd wewnętrzny serwera", http.StatusInternalServerError)
		}
		return
	}

	if err := bcrypt.CompareHashAndPassword([]byte(storedHash), []byte(req.Password)); err != nil {
		http.Error(w, "Nieprawidłowe dane logowania", http.StatusUnauthorized)
		return
	}

	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
		"username": req.Username,
		"exp":      time.Now().Add(time.Hour * 24).Unix(),
	})

	tokenString, err := token.SignedString(jwtSecretKey)
	if err != nil {
		http.Error(w, "Błąd podczas generowania tokena autoryzacji", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	err = json.NewEncoder(w).Encode(map[string]string{"token": tokenString})
	if err != nil {
		return
	}
}